Basic game created.
Basic gameplay elements are mostly present. Tile selection (both player and random computer selection) is implemented. Player looking and environment also added. Performance is fine right now but some code is sloppy and needs to be refactored. For instance, the random tile selection is not optimal and would benefit from an integer array of tile statuses so it doesn't have to dig deep into the tiles' encapuslated items. UI elements, endgame, scoring, timer and menu still need to be implemented.