Started main menu.
Sketched a title and added a menu scene. Start button added with script that starts the levels scene and essentially starts the game. Leaderboard, exit and help buttons need to be added. Game over screen still needs to be added. It should display the final score, level and placement in the leaderboard.